+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Profile header rule now strictly requires full semver version in version field, as the spec defines
+- Condition to add SingleLineCommentTrivia to termina tokens in sublexer
 
 ## [2.0.0] - 2022-11-08
 ### Added