+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.1.0] - 2023-01-31
 ### Added
 - Added syntax rule for `None` literal
 
@@ -251,7 +253,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Documentation extraction from doc strings
 - Usecase result parsing as optional
 
-[Unreleased]: https://github.com/superfaceai/parser/compare/v2.0.0...HEAD
+[Unreleased]: https://github.com/superfaceai/parser/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/superfaceai/parser/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/superfaceai/parser/compare/v1.2.0...v2.0.0
 [1.2.0]: https://github.com/superfaceai/parser/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/superfaceai/parser/compare/v1.0.0...v1.1.0